packages/zchat-protocol/pyproject.toml
[project]
name = "zchat-protocol"
version = "0.0.6.3"
description = "ZChat protocol primitives — type definitions only"
requires-python = ">=3.12"
dependencies = ["python-ulid>=3.0", "typing_extensions>=4.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zchat_protocol"]
